requirements.txt
sentence-transformers
streamlit
numpy
pandas
scikit-learn
google-generativeai
plotly
